remove spyrmsd module and add __main__ module
