+++ README.md
@@ -73,7 +73,7 @@ The following packages are required to use `spyrmsd` as a module:
 * [scipy](https://www.scipy.org/)
 * [qcelemental](http://docs.qcarchive.molssi.org/projects/qcelemental/en/latest/)
 
-_Note_: `spyrmsd` uses [graph-tool](https://graph-tool.skewed.de/) by default but will fall back to [NetworkX](https://networkx.github.io/) if the former is not installed (e.g. on Windows). However, in order to support cross-platform installation [NetworkX](https://networkx.github.io/) is installed by default.
+_Note_: `spyrmsd` uses [graph-tool](https://graph-tool.skewed.de/) by default but will fall back to [NetworkX](https://networkx.github.io/) if the former is not installed (e.g. on Windows). However, in order to support cross-platform installation [NetworkX](https://networkx.github.io/) is installed by default, and [graph-tool](https://graph-tool.skewed.de/) needs to be installed manually.
 
 #### Standalone Tool
 
@@ -89,11 +89,11 @@ _Note_: [RDKit](https://rdkit.org/) is not available on PyPI ([Why the RDKit isn
 ### Standalone
 
 ```bash
-python -m spyrmsd.spyrmsd -h
+python -m spyrmsd -h
 ```
 
 ```text
-usage: spyrmsd.py [-h] [-m] [-c] [--hydrogens] [-n] reference molecules [molecules ...]
+usage: __main__.py [-h] [-m] [-c] [--hydrogens] [-n] reference molecules [molecules ...]
 
 Python RMSD tool.
