+++ spyrmsd/molecule.py
@@ -235,3 +235,33 @@ class Molecule:
             Number of atoms within the molecule
         """
         return self.natoms
+
+
+def coords_from_molecule(mol: Molecule, center: bool = False) -> np.ndarray:
+    """
+    Atomic coordinates from molecule.
+
+    Parameters
+    ----------
+    mol: molecule.Molecule
+        Molecule
+    center: bool
+        Center flag
+
+    Returns
+    -------
+    np.ndarray
+        Atomic coordinates (possibly centred)
+
+    Notes
+    -----
+    Atomic coordinates are centred according to the center of geometry, not the center
+    of mass.
+    """
+
+    if center:
+        coords = mol.coordinates - mol.center_of_geometry()
+    else:
+        coords = mol.coordinates
+
+    return coords
